Add delete_item to Widgets Controller
Added a delete_item method to the widgets controller.  I wasn't sure
what it should return so it currently just returns the settings so when
all widget instances of a type are deleted it returns an empty array.
Suggestions on what it should return on success would help.  This should
work with #17 and #18 should also work with #17.